+++ src/CogServer/CogServer.cc
@@ -102,7 +102,7 @@ void CogServer::serverLoop()
         elapsed_time = ((timer_end.tv_sec - timer_start.tv_sec) * 1000000) +
                        (timer_end.tv_usec - timer_start.tv_usec);
         if ((cycle_duration - elapsed_time) > 0)
-            usleep(cycle_duration - elapsed_time);
+            usleep((unsigned int) (cycle_duration - elapsed_time));
         timer_start = timer_end;
     }
 }
