+++ src/nlp/wsd/SenseRank.cc
@@ -161,7 +161,7 @@ void SenseRank::rank_sense(Handle h)
 	converge *= (1.0-convergence_damper);
 	converge += convergence_damper * fabs(rank_sum - old_rank);
 
-	SimpleTruthValue stv(rank_sum, 1.0);
+	SimpleTruthValue stv((float) rank_sum, 1.0f);
 	stv.setConfidence(sense->getTruthValue().getConfidence());
 	sense->setTruthValue(stv);
 }
